+++ src/redturtle/volto/tests/test_summary_customization.py
@@ -9,7 +9,9 @@ from plone.app.testing import (
 from plone.restapi.testing import RelativeSession
 from plone.namedfile.file import NamedBlobImage
 from redturtle.volto.testing import REDTURTLE_VOLTO_API_FUNCTIONAL_TESTING
-
+from datetime import datetime
+from plone.app.querystring import queryparser
+from plone.restapi.deserializer import json_body
 import unittest
 import transaction
 import os
@@ -64,7 +66,9 @@ class TestSummaryCustomization(unittest.TestCase):
 
         self.assertEqual(len(res["items"]), 2)
         self.assertEqual(res["items"][0]["title"], self.news_with_image.title)
-        self.assertEqual(res["items"][1]["title"], self.news_without_image.title)
+        self.assertEqual(
+            res["items"][1]["title"], self.news_without_image.title
+        )
         self.assertNotIn("image", res["items"][0])
         self.assertNotIn("image", res["items"][1])
 
@@ -75,15 +79,15 @@ class TestSummaryCustomization(unittest.TestCase):
 
         self.assertEqual(len(res["items"]), 2)
         self.assertEqual(res["items"][0]["title"], self.news_with_image.title)
-        self.assertEqual(res["items"][1]["title"], self.news_without_image.title)
+        self.assertEqual(
+            res["items"][1]["title"], self.news_without_image.title
+        )
         self.assertIn("image", res["items"][0])
         self.assertNotIn("image", res["items"][1])
 
     def test_summary_return_empty_effective_date_if_not_set(self):
         page = api.content.create(
-            container=self.portal,
-            type="Document",
-            title="Document",
+            container=self.portal, type="Document", title="Document",
         )
         transaction.commit()
 
@@ -105,3 +109,37 @@ class TestSummaryCustomization(unittest.TestCase):
 
         self.assertEqual(len(res["items"]), 1)
         self.assertEqual(res["items"][0]["effective"], page.effective_date)
+
+    def test_event_summary_on_recurrence(self):
+        event = api.content.create(
+            container=self.portal, type="Event", title="Event with recurrence",
+        )
+        event.start = datetime(2022, 1, 9, 13, 0)
+        event.end = datetime(2022, 4, 24, 14, 0)
+        event.recurrence = "DTSTART:20220109T000000Z\nRRULE:FREQ=WEEKLY;INTERVAL=1;UNTIL=20220424T000000Z;BYDAY=SU,SA"  # noqa
+        event.reindexObject()
+        transaction.commit()
+        res = self.api_session.post(
+            self.portal_url + "/@querystring-search",
+            json={
+                "metadata_fields": "_all",
+                "b_size": 24,
+                "query": [
+                    {
+                        "i": "portal_type",
+                        "o": "plone.app.querystring.operation.selection.any",
+                        "v": ["Event"],
+                    },
+                    {
+                        "i": "start",
+                        "o": "plone.app.querystring.operation.date.afterToday",
+                        "v": "",
+                    },
+                ],
+                "sort_on": "start",
+                "sort_order": "ascending",
+                "b_start": 0,
+            },
+        )
+        item = res.json()["items"]
+        self.assertEqual(len(item), 0)